+++ cockpit/src/main/java/com/polidea/cockpit/paramsedition/viewholder/GroupViewHolder.kt
@@ -2,20 +2,16 @@ package com.polidea.cockpit.paramsedition.viewholder
 
 import android.view.View
 import android.widget.TextView
+import android.widget.Toast
 import androidx.recyclerview.widget.RecyclerView
 import com.polidea.cockpit.R
 
 internal class GroupViewHolder(val view: View) : RecyclerView.ViewHolder(view) {
 
-    private val separator: View = view.findViewById(R.id.cockpit_group_separator)
-    private val name: TextView = view.findViewById(R.id.cockpit_group_name)
+    private val name: TextView = view.findViewById(R.id.subgroup_name)
 
     fun display(groupName: String?) {
-        if (adapterPosition == 0)
-            separator.visibility = View.GONE
-        else
-            separator.visibility = View.VISIBLE
-
         name.text = groupName ?: view.context.getString(R.string.default_group_name)
+        view.setOnClickListener { Toast.makeText(view.context, groupName, Toast.LENGTH_SHORT).show() }
     }
 }
